add testcase for subscriber
